Remove Build Rule events from WebSocket to improve event bus perf.
Summary:
During build of heavy apps, BuildRule events generate a lot of traffic to the WebSocket.
This means it slows down the whole buck event bus, so both WebSocket and Superconsole are far away from
current state of the build.
By reducing number of events we post into WebSocket, we reduce the traffic and make Event Bus faster.
Also, since we already have progress events reported to the WebSocket, we should use them instead of
trying to calculate the progress based on the build rule event sequence.

Test Plan: check that events do not appear in WebSocket interface during the build.