Make HttpDownloader not throw on error
Summary:
The recent change "Introduce a stacked Downloader" made it possible for
multiple Maven repositories to be available during the build, but it
wasn't possible in practice to use them because the first repository
would throw an exception and the others would not be checked.  Change
from an exception to a log message and "return false".

Test Plan:
Built an app that required "git fetch" to get libraries from both Maven
Central and the local Android and Google m2 repositories.