java_test: don't cache classpath argsfile using library rule key
Summary:
There were a few issues caching this:
1) Since `JavaTest` is a child of `DefaultJavaLibrary`, it was using
   the latter's rule-key and input-based-rule-key.  This is a problem
   since the rule keys for `DefaultJavaLibrary` are explicitly non-
   transitive, whereas the classpath argsfile includes the all the transitive
   libs.  This means we could get a cache hit via the input-based-rule-key
   and pull down an argsfile that doesn't take into account updates to
   the transitive classpath.
2) The paths the JARs under the android SDK are absolute, so we don't
   want to cache these.  Also, with cross-cell support incoming, that'll
   likely be another source of absolute paths.

Since generating the classpath file is relatively cheap, this diff changes
`JavaTest` to use post-build-steps instead, which run even on cache hits,
meaning we don't have to cache the classpath argsfile.

Test Plan:
Ran a `java_test`, changed a transitive dependency and verified it was
reflected in the classpath argsfile, even when we get a cache hit.