Add resource_union_package for android_library
Summary:
When Gradle-using apps generate the Uber R.java with aapt, it seems that
aapt will put *all* resource ids into the main package.  This seems kind
of unsafe to me, but some apps depend on it.  Ideally, we'd get this
behavior for free by using aapt to generate the Uber R.java for the
"application sources", but for now, here's a hack to request this
behavior explicitly for an android_library.

I'm deliberately not documenting it because I don't think we should
commit to supporting it long-term.

Test Plan:
Successfully built an app that was depending on that behavior.
Unit test.