Set DEVELOPER_DIR environment variable when running tests.
Summary:
We currently don't always set DEVELOPER_DIR when invoking xctool, xctest
and octest. This means even though we build the tests with the configured
version of Xcode, we will still run them using the version of Xcode returned
by xcode-select or configured using DEVELOPER_DIR. This fixes that.

Test Plan:
CI
Configure a version of Xcode with xcode-select, then run a test by overriding it
in .buckconfig. Verify it is run with the right version of Xcode.