Push the knowledge of the TargetNode hash into the TargetNode.
Summary:
The way that target key calculation used to work was by hoping
that the caches within the `Parser` had been warmed so that by
coincidence the values required existed when accessed. It goes
without saying that this isn't a great thing to do.

We now store the hash of the JSON used to create `TargetNode`s
on the node itself, which avoids this problem.

Test Plan: buck test --all