Use a stream to write Dalvik-aware dex splitting report
Summary:
This is similar to "Use a stream to write dex splitting report", except
that it fixes the case when `use_linear_alloc_split_dex` is set to
`True` on an `android_binary` rule.

The `MorePaths.append` method would open, append, flush, and close the
report file for every class in the resulting `.apk`, which resulted in
dex splitting being very IO-bound.

Test Plan: CI