[WS] main changes to the plugin
Summary:
These are the main changes to the plugin.
- actions are disabled if there is no connection to buck.
- buck actions (build, kill etc) are still issued in the old fashion.
- all the output comes from a websocket connection.
- errors are now displayed in the "Build" buck tool window tab. They are grouped by target/file (you need to run buck with --keep-going, which is disabled right now in the plugin, to get multiple errors). Error are clickable.

Known "issue": because the way buck works internally, if a high level target is found in cache, the deps are not marked as finished, the messages BuildRuleFinished and BuildRuleResumes are not reliable. The UI for total tasks, running, finished and suspended is informative only (you might get a 1000 total targets, 999 suspended, 1 finished - because buck found the top most target in cache and marked only that as completed). Also, because the cli takes some time after build is finished to return, the progress bar will show for a bit more after the build is finished.

Test Plan: Build and install the plugin OR run from intellij. Open a project. Build. Open a project and deliberately break some files. Use buck build and shortly you should see the errors.