Support for flavor propagation across apple_bundle dependents
Summary:
Right now, when we build a bundle A that depends on another bundle B, we don't pass A's platform flavor (so we fall back to the default flavor when building B).  Fix this so we properly do this.

This is needed for Watch apps, that consist of a `.appex` bundle embedded within a `.app` bundle.

Test Plan: CI