Rewrite the Parser with cross-cell caching in mind.
Summary:
An reimplementation of the original Parser, but ready to
handle cross-cell dependencies, and with state being
more cleanly managed.

Essentially, any operation that will cause state to be loaded
must go through the `PerBuildState`. This, in turn, will call
down to the `DaemonizedState` to obtain the actual values.

Test Plan: * Migrate the tests from `ParserTest` to `ParserNgTest` and make them pass.