Easy: Fix bugs in TestDescription.equals()
Summary:
I'm adding support to the Apple xctool test runner for test filtering
via `buck test -f`.

When I did this, I found `TestDescription` (which holds the result of
parsing the parguments passed to `buck test -f`) had a buggy
`equals()` method: it assumed that strings were equal if their hashes
were equal.

I went to replace this with an `Immutables.org` class, but this class
is used by the JUnit test runner, which is compiled very gingerly and
without support for Immutables annotation processing.

Since that looked annoying to fix, I just fixed the `equals()` method and
added support for `toString()` as well.

Test Plan: Used in subsequent diff.