Rename methods in the SourceResolver and allow relative paths for BuildTargetSourcePath
Summary:
The `getPath` method is a confused beast, in that it sometimes
returns an absolute and sometimes returns a relative path when
called. There is a widespread assumption in the tree that this
always returns a relative path. Fix that assumption.

This is the first in a set of diffs to make Buck function well
with cross-cell diff and file name resolution.

Test Plan:
Added a test
buck test --all