Remember Watchman warnings encountered during query so we can pass them to Watchman glob handler
Summary:
Previously, if Watchman posted a sticky warning, we would
show it both when `WatchmanWatcher` encountered it and when
the Watchman glob engine encountered it.

This adds logic to remember warnings emitted by Watchman so
in a subsequent diff I can pass it to the parser so we can avoid
printing them twice.

Test Plan: Unit tests updated. `ant java-test -Dtest.class=WatchmanWatcherTest`