[halide] Build "universal" binaries for Halide libraries
Summary:
Update the halide compiler script template used in Xcode project generation to
support building for multiple architectures (e.g. armv7 and arm64) and creating
a "universal" binary from the resulting object files.

Test Plan:
Ran ##buck project## on a project that uses a ##halide_library## target, opened
the resulting project in Xcode and built and ran the application on an iPhone5,
iPhone5s, and in the simulator.