Send BUCK file parser warnings and errors from Python back up to Java
Summary:
Previously, the protocol between `buck.py` and Java was very
simple: either it barfed errors to stderr, or it serialized the
parsed `BUCK` file contents to stdout.

We want to be able to filter certain errors and warnings (say, when
Watchman repeatedly warns about an issue), so this diff updates
the protocol between `buck.py` and Java so we can pass warnings
and errors back up.

Previously, the protocol was that `buck.py` wrote a map of
parsed `BUCK` file contents to stdout.

Now, the protocol is that `buck.py` will write a map with
one or two keys:

1) `values`: The parsed `BUCK` file contents
2) `diagnostics` (optional): A list of warnings and errors encountered during parsing (currently maps of `level` and `message`, later we can make these richer)

Later, we can extend this to serialize Python exceptions and
send them up to Java so we can build much better UI to indicate
typos in `BUCK` files.

Test Plan: Unit tests included. `ant java-test -Dtest.class=ProjectBuildFileParserTest`