Fix handling of different resource types with same name
Summary: This code mixed resource names for different resource types (strings, plurals, string-arrays), despite the fact that Android only enforces uniqueness of names within a resource type. This caused overflow bugs when we tried to build .fbstr files for an app that has a <plurals> and <string> element with the same name in different modules. This changes the behaviour to be consistent with Android resource name handling. Additionally, output of GenStringSourceMapStep is now grouped by resource type which is consistent with the documentation for that step that was added previously.

Test Plan: Tests, see that building the app that was broken previously now works, look at string source map and see output is grouped by resource type.