Add a hang detector.
Summary:
This will occasionally dump out the stacks of all of the threads
in the JVM process to the log file. The stacks are intentionally formatted
to appear in the log as a single line so that they don't attract the attention
of anyone looking for crash-related stack traces and that they don't take
up too much space.

Test Plan:
CI, also manually verified this is WAI by adding a sleep statement
in the main loop and observing the thread dump go into the logs.