[halide] Get halide target string from rule flavor
Summary:
As title. This makes it possible to build halide_library rules for a device
(e.g. iPhone5s) from the command line, by doing something like:

  buck build path/to/my/lib:lib_target#iphoneos-arm64

And have the Halide compiler be invoked with the correct target string (in this
case, "arm-64-ios") for the given target architecture and platform.

Test Plan:
Built, installed, and launched an application which depends transitively on a
halide_library target using buck, both in the simulator and on an iPhone5s.