Hook up android_instrumentation_test to test infrastructure
Summary:
When android_instrumentation_test was introduced, an integration test
was added, but I forgot to hook it up to the test infrastructure.

This caused us to miss that some binary files broke when importing the
pull request from GitHub.

Hook up the integration test to the test infrastructure and fix the
broken binary files.
Closes https://github.com/facebook/buck/pull/507

Test Plan:  buck test with and without android devices attached