Pure Python Nailgun client for buckd
Summary:
We've constantly struggled with building the native Nailgun C
client as part of distributing Buck.

Adding the requirement of a C compiler is a lot to ask for people
building pure Java projects. We often run into users who have to deal
with things like accepting Xcode license agreements as root just to
build Buck, which seems silly.

This diff adds a pure Python port of the Nailgun client so we can
distribute a pure Java distribution of Buck that works with `buckd`.

I made sure the Python code worked both with Python 2.6 and 2.7.

This doesn't yet work on Windows, but we don't support `buckd`
on Windows yet, so it's not a net loss of functionality.

I sent this pull request upstream as well:

https://github.com/martylamb/nailgun/pull/80

Test Plan:
Tested on Linux 3.10.75 with Python 2.6 and OS X 10.10.4 with Python 2.7.

Started `buckd` with `buck build buck`. Connected Python
client and built again by running:

  python third-party/nailgun/nailgun-client/ng.py --nailgun-server=local:.buckd/sock com.facebook.buck.cli.Main build buck