Delete TestResultsGrouper and deprecate `buck test --ignore-failing-dependencies`
Summary:
I'm trying to make `TestResults` an immutable value type.

When I went to do so, I ran into `TestResultsGrouper`, which mutates (!)
`TestResults` objects and appears to rely on the side effects of
leaving the mutated objects in that object's cache.

The only place `TestResultsGrouper` is used is for the `buck test`
command-line option `--ignore-failing-dependencies`. Nobody appears to
use that, so this diff deprecates that option and makes it a no-op,
then deletes `TestResultsGrouper` and its many tests.

Test Plan: CI