Fix loading libraries with sonames from Java on Mac
Summary:
It seems that System.loadLibrary on Mac will only load libraries named
as "*.dylib".  However, we have a library used by a Java test that has
an explicit soname set to "libsomething.so".  One option is for us to
use the default soname, which automatically gets the proper .so/.dylib
suffix, but that requires hard-coding that long name in the Java
wrapper, which is sub-optimal.

This diff gives us another option.  The $(ext) macro will be replaced
with the appropriate extension.  An optional version number will be
formatted appropriately as well.

Closes https://github.com/facebook/buck/issues/339

Test Plan:
Integration test.  I made it use an explicit soname, and added another
library to continue testing the non-soname case.

Viewed docs.