Revert "Extend depfile rule keys to support added headers."
Summary: It doesn't appear that this change actually did anything.  The
`CxxHeaders.appendToRuleKey` method still attempts to set all
it's included headers, which we'd we need to change to actually
do this.

I've left the test it added the reverted diff added to verify
we still have the desired behavior.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: 0e6977a