python: have in-place and standalone use same rule for executable
Summary: Previously, the in-place python binary rule actually used a `WriteFile`
rule with a `#script` flavor to generate the output.  This meant if that
if the user was switching between the in-place and standalone package
modes, that they'd run two different build rules outputting to the same
location (e.g. `//foo:bar` for standalone would output to `buck-out/gen/foo/bar.par`,
while `//foo:bar#script` for the in-place variant would also output to
`buck-out/gen/foo/bar.par`).

In general, this can't work, since the build engine just checks the
rule key metadata files to see if a rule needs to run, and so it will
never know that a previous run with a standalone binary overwrote the
output for an in-place binary, leading to bugs with stale outputs.

This diff changes the in-place variant to write the output file using
the same build target as the standalone variant to avoid this issue.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: 5200ea5