Fix Python launcher failing to use buckd if run outside of project root
Summary: I found some engineers were running buck from directories outside
the root of the repository.

In this case, the Python launcher would fail to connect to `buckd`, because
we were passing it a relative path to the Unix domain socket which was
hard-coded to `.buckd/sock`.

This fixes the issue by updating `pynailgun` to support a `cwd` parameter
to tell it to chdir to the project root (and then restore it) before
connecting to the socket.

We could pass the full absolute path to the socket, but Unix domain
sockets are extremely limited in length (somewhere around 60 bytes)
so using a relative path if possible is much more reliable.

Test Plan: `cd` into subdirectory, run Buck command twice. Before,
would not use buckd. Now, it does.

Reviewed By: sdwilsh

fb-gh-sync-id: 5095ebd