Add debug logging to Nailgun server and potential fix for client failure
Summary: I'm getting reports from users that the Nailgun Python client
is reporting the server is unexpectedly disconnecting the socket connection.

This adds the following debug logging:

https://github.com/bhamiltoncx/nailgun/commit/c2788d68d67247a1b30ef7a857027f9b877827ea

so we can get information on the Buck side about what's
happening. (From the client's perspective, all it sees is that the
server is disconnecting.)

It also includes an extra flush in `NGInputStream` as a guess as to what is causing the problem.

Test Plan: buck build buck, confirm Nailgun logs in buck-out/log/buck-1.log
and buck-out/log/buck-0.log

Reviewed By: Coneko

fb-gh-sync-id: 5ad9514