PackagedResource: make unpacked resources persistent between runs
Summary: Previously, the resources that the `PackagedResource` class unpacked
were cleaned up when Buck exited.  If the resource was used in a
symlink tree, this meant that when not using buckd, the symlink tree
would be left with a dangling link to the unpacked resource.

This diff switches the `PackagedResource` to use a persistent unpack
dir, under `buck-out/res/<name>`.  When resolved, the resource is
always freshly unpacked the at most once during a given Buck run.

Test Plan: - CI
- Ran a buck build without buckd and verified resource still existed.

Reviewed By: alisdair04

fb-gh-sync-id: 2f7dfaa