Speed up parser's handling of cached targets
Summary: After switching from `Parser` to `ParserNg`, the parse times for cached
targets regressed because `ParserNg` used maps with `Cell` as keys which
are fairly expensive to hash and compare. To speed it up, use only the
cells' roots as keys. Also, since cell's equality takes into account the
value of its `BuckConfig`, it was actually impossible to find a value in
the map that would be distinct from the expected one.

Test Plan: `buck test`. Compare performance on repeated runs of
`buck targets --show-output`, notice significant speedup.

Reviewed By: bhamiltoncx

fb-gh-sync-id: 1e38005