Remove usage of deprecatedPathFunction from OCamlRuleBuilder
Summary:
The OCaml rule compiles dependent .c files by flattening their paths
(just takes the last component), resolving them against an out dir
and sticking the .o files there. We'll fail to create the ActionGraph
if multiple files resolve to the same name (as was the case before
this patch). It doesn't really matter which path function we use
(as the code that consumes the Path only looks at the file name)
so using absolute paths to be on the safe side.

Test Plan: CI

Reviewed By: k21

fb-gh-sync-id: d2d379f