Fix dSYM path generation for bundles.
Summary:
For OS X style bundles, the executable is in a nested directory, so we can't just request the parent of the executable path. Instead, use the bundle's parent path, which is always in the right place.

Additionally, use the bundle name, not the executable name, for the name of the dSYM.

Test Plan: Tested by building a sample app with the nested bundle style. Existing integration tests verify correct behavior for the flat bundle style.

Reviewed By: beefon, ryu2

fb-gh-sync-id: 84e46e8