Remove findBuildRuleByTarget.
Summary:
With the new uses of the require pattern a description's `createBuildRule`
method doesn't always return a rule with the same build target as was requested.
This means the resolver could end up mapping the same build rule to different
build targets. This causes `findBuildRuleByTarget` to break. Luckily all call sites
of that method have easy ways of accessing the resolver used to create the
action graph, so they can use require instead.

Test Plan: CI

Reviewed By: @beefon

fb-gh-sync-id: 72eb597