[halide] Add halide library to "Link Binary" phase of Xcode target
Summary:
Instead of adding the archive produced by the halide_library target to the
EXTRA_LD_FLAGS, add a reference to the file to the "Link Binary" phase of
targets which depend on it. This is for consistency with how Buck handles
other rules (e.g. cxx_library), and it also allows users to use alternate
versions of the library by changing the linker search paths.

Test Plan:
Did 'buck project' for a project which depends transitively on a halide_library.
Opened the project in Xcode and confirmed that the library was included in the
"Link Binary" phase of the application target. Built the project in Xcode and
made sure that it linked and ran correctly.

Reviewed By: Coneko

fb-gh-sync-id: 0183364