Delete test utility SuperConsoleEventBusListenerTest.formatConsoleTimes()
Summary:
This is a follow-up to my previous diff
`Fix SuperConsoleEventBusListenerTest when run with non-English locales`.

In this diff, I delete `SuperConsoleEventBusListenerTest.formatConsoleTimes()`,
since we no longer need that workaround.

Previously, we would try to change the tests based on the locale in
which they were running. Now, we pass the `Locale` down to `SuperConsoleEventBusListener`,
so we can explicitly test whatever locale we want.

Test Plan: `ant java-test -Dtest.class=SuperConsoleEventBusListenerTest`

Reviewed By: sdwilsh

fb-gh-sync-id: 0eb023b