Add SuppressForbidden annotation and use in AdbHelper
Summary:
To make progress enabling `ant forbidden-checks` as part of
the build, we need a way to suppress the checks on legacy code and/or
code we know is good.

This adds a new annotation `com.facebook.buck.annotations.SuppressForbidden`
and enables its use from `build.xml` to allow us to turn off the warnings.

I added calls to `AdbHelper` to turn off the warning for existing unsafe
calls to `String.format()` as an example. (We could also pass a `Locale`
into this class if we wanted to fix the errors more properly.)

Technically we could add `forbiddenapis-1.8.jar` to Buck's classpath to
get their copy of `SuppressForbidden`, but we're not using it today,
so I added our own annotation instead to avoid the dependency.

Test Plan:
`ant forbidden-checks`. Before, `AdbHelper` appeared many
times. Now, it doesn't.

Reviewed By: k21

fb-gh-sync-id: 442196e