Stop creating an extra target graph in `buck targets --show-target-hash`
Summary:
Probably through a series of refactorings, we've managed to create
an extra copy of the target graph.  This change removes this madness.

The use of `ProjectGraphParser` was added at some point to make it
easier to write tests for this, however those tests have not
materialized and the abstraction likely led to a series of refactors
that managed to introduce this extra copy because it was non-obvious
that the graph created from the `ProjectGraphParser` was identical to
the target graph.

Test Plan: CI

Reviewed By: bhamiltoncx

fb-gh-sync-id: 3cc6955