New method ProjectFilesystem.isSameRelativePathIfFileExists()
Summary:
We have a number of places which assume that `String.compareToIgnoreCase()`
or `Path.equals()` works correctly to match paths in a case-insensitive manner.

The actual rules depend on the filesystem (and may require doing disk
I/O), so we have to use `Files.isSameFile()` to ask the filesystem if
the two paths are the same.

This diff introduces a new method
`ProjectFilesystem.isSameRelativePathIfFileExists()` which takes two
project filesystem-relative paths and asks the underlying filesystem
if they refer to the same file on disk.

If either path does not exist, this method returns `false`.

Test Plan: Unit tests included. `ant java-test -Dtest.class=ProjectFilesystemTest`

Reviewed By: ryu2

fb-gh-sync-id: f4ea2b9