Fix Turkish I bugs in AndroidResourceDescription
Summary:
`AndroidResourceDescription` tries to filter out
various paths under the `res/` directory based on the filename,
directory name, or suffix.

Unfortunately, it uses `String.toLowerCase()` to try to do
a case-insensitive match on the filename. This doesn't work
correctly in languages like Turkish, since the case-mapping
rules there for the letter "i" are not the same as English.

This fixes the bug by using the new utility
`ProjectFilesystem.isSameRelativePathIfFileExists()` to ask
the underlying filesystem whether the path is the same.

Test Plan: Unit tests updated. `ant java-test -Dtest.class=AndroidResourceDescriptionTest`

Reviewed By: ryu2

fb-gh-sync-id: 2a86448