[halide] Make HalideLibrary cache key depend on target map
Summary:
Changing the platform to target map in .buckconfig should invalidate any cached
results for halide_library targets, since the generated code depends on values
in the map. This change adds the contents of the map to the rule key for
HalideLibrary rules.

Test Plan:
Built a halide_library target, then changed the Halide target mapping in my
.buckconfig. Built the same target again and confirmed that the 'compiler' for
the pipeline was re-run with the new target string. Inspected the generated
header file for my pipeline after both runs to confirm it matched what I
expected.

Reviewed By: grp

fb-gh-sync-id: d7610c5