Account for additional entitlement keys when matching provisioning profiles
Summary:
When choosing a provisioning profile for iOS device builds, consider the additional keys in the entitlements file (if supplied).

Otherwise, we could still be potentially signing with a provisioning profile that doesn't have the necessary entitlements for the app, causing an error when installing to device.

Test Plan: CI

Reviewed By: yiding

fb-gh-sync-id: 9fb83fe