Allow the specification of an agentpath for the invocation of ProGuard
Summary: To help with profiling it would be useful to run ProGuard with an -agentpath argument. This diff allows the specification of the agentpath in the [tools] section of .buckconfig or .buckconfig.local

Test Plan: Run "ant clean test" and run builds of FB4A with and without profiling enabled for ProGuard.

Reviewed By: bhamiltoncx

fb-gh-sync-id: 55a3a65