Java 8 requires all interfaces to be present when compiling.
Summary:
Interfaces need to be present when compiling against their
implementations.

When compiling a class against another class implementing
an interface which is defined in yet another class file,
such class file (where interface is defined) must be available
in the class path used by javac during compilation. This
is a new requirement as of JDK 8 - a failure to do so will
result in a compilation error.

Source: http://www.oracle.com/technetwork/java/javase/8-compatibility-guide-2156366.html

This was being seen when building the Apple rules.

Test Plan: `buck build buck` on java 8

Reviewed By: Coneko

fb-gh-sync-id: a1fff5c