Don't restart the daemon when the number of threads is changed.
Summary:
Buck restarts the daemon on every config change. While this is correct
it also means we re-parse the graph when the number of threads changes
or when the cache timeouts are updated (for example).

Test Plan:
Compare `buck targets --show-output` with
`buck targets --show-output --config=build.threads=1`, observe no parsing on
second invocation.

Reviewed By: shs96c

fb-gh-sync-id: 8e5236c