Add the selected target to the rulekey instead of the whole target map.
Summary:
We currently add the target map to the rule key. This means the rule key
does not change if the selected target changes, and does change if the
target map changes in a way that doesn't affect the selected target.
This changes the rule key to take the selected target into account and
not the target map.

Test Plan:
CI
Build a halide target, add a new halide target to the map, build it again

Reviewed By: k21, jcarreiro

fb-gh-sync-id: dd38b92