Allow setting a target build-tools version
Summary:
This change resolves build tools in directory resolver instead of the
`AndroidPlatformTarget` class in order to allow us to enforce that we
have found a specific version of the build tools for Android.

A future change can build off of this to then make tools like `aapt`
a `VersionedTool` that can be added to the rule key to make our
builds more correct.

A nice side effect of this change is that the complicated logic on
what directory to pick now is easier to test and the tests are now
much simpler.

Test Plan:
* added/ported tests for the cases that are new for the directory resolver
* existing tests
* `./docs/soyweb-local.ps1` and loaded the buckconfig page

Reviewed By: marcinkosiba

fb-gh-sync-id: d9351e3