[ios] Only select certificates that can be signed by the available identities
Summary:
Make `ProvisioningProfileStore` optionally match only certificates which are signable by a given set of identities.

This prevents the scenario where for instance:

1. A developer's signing certificate expires and is renewed
2. The associated profiles regenerated with the renewed certificate and redownloaded from Apple, but the old profiles with the old certificate are still around on disk.

In this case, we would potentially pick the old profiles which don't have the current certificate instead of the newer one, and therefore be unable to sign later on.

Test Plan: CI

Reviewed By: yiding

fb-gh-sync-id: 6d49b2c