Fix ANSI codes found in Go tool output
Summary: When trying to use the go tool output to discover the location of the linker and compiler there was ANSI color codes that were being leaked in causing the output to the wrong and fail.  This sets the ProcessExecutor to expect STD_OUT which solves the issue.

Test Plan: Compiled a local go_binary without any specific [go] settings in .buckconfig

Reviewed By: Coneko

fb-gh-sync-id: 8ce3c32