Enable the dir cache by default
Summary:
Lots of folks fail to setup the directory cache when they start out with Buck.  As
a result, they aren't getting all the benefits.  This change enables the cache by
default under `buck-out/cache` so that nuking `buck-out` will also clear the
cache.

To disable this, you can add this to your `.buckconfig`:
```
[cache]
  mode =
```

Test Plan: CI, which happens to catch a lot of issues with the simple approach.

Reviewed By: bolinfest

fb-gh-sync-id: cdc9ade