[perf] Allow generating temp file paths without creating empty files on disk to minimise IO.
Summary:
We call Java's methods to obtain the path to temp file, but we don't need that file to be created.
This diff optimizes this so now there is a way to get a path without creating a file on disk.
This minimizes the disk I/O a bit, speeding up local cache check operation mostly.

Test Plan: CI

Reviewed By: marcinkosiba

fb-gh-sync-id: 383212e