[perf] Do not create / delete manifest prior to checking cache
Summary:
In attempt to lessen disk I/O, we can avoid unnecessary operations like deleting manifest from disk in case if it does not exist in cache.
Also, we avoid preparing the hierarchy of folders in case of cache miss. We prepare it only if cache hit happens as well as if we build the manifest locally.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: a3436d0