Pass down all immediate linker options instead of just frameworks
Summary:
System libraries required by shared libraries directly (as opposed to via
dependent libraries) were not being properly passed down, leading to unresolved
symbols when building libraries.

This fixes that, and consolidates the trifecta of frameworks, libraries, and
linker args that's coming from the shared library itself into a
`NativeLinkableInput`, and treated the same as the ones retrieved from the
dependent libraries.

A similar change for binaries is in a subsequent diff.

Also fix libraries to library search path mapping to not strip an extra layer
of directories.

Test Plan:
Added unit tests for this case.

Built a library that has an entry in `libraries` with '#shared'. See that the
link succeeds.

Reviewed By: Coneko

fb-gh-sync-id: bad17ff