omnibus: don't include unused always-static libs in libomnibus
Summary:
This fixes a minor bug where a library requesting static linkage
would get included in the merged libomnibus library even if no
other libs in it used it.  This could happen if a root had a dep
which marked itself as `link_whole=True`, and no other libs used
it.  In this case, we should statically link it into the root and
leave it out of libomnibus.

Test Plan: added unittest

Reviewed By: sdwilsh

fb-gh-sync-id: 350ded0