Remove fields from Apple*Description that are derivable from delegate
Summary:
This removes potential a degree of freedom that allows cxxPlatform to differ
from that in delegate, improving the clarity of the code.

Also changes how `createBuildRule` is invoked, namely `getTypeAndPlatform` no
longer needs to be invoked in `AppleLibraryDescription`, since `cxxPlatform`s
are guaranteed to be the same.

Test Plan: unit test passes

Reviewed By: andrewjcg

fb-gh-sync-id: 0069b66
shipit-source-id: 0069b66