Make CxxSourceRuleFactory PicType parameter a field
Summary:
`PicType` is a property on the library and doesn't vary within source files of a
library.  Therefore it ought to be a parameter of the factory itself, not
constituent functions.

The only cases where a different `PicType` is passed to the same factory is in
tests, which are fixed by simply creating two `CxxSourceRuleFactory`s.

Test Plan: unit test passes

Reviewed By: Coneko

fb-gh-sync-id: 7d5812b
shipit-source-id: 7d5812b