Easy: Forbid symlinks to source files in Buck repo
Summary:
Buck's Parser and Watchman have a number of difficulties
with symlinks to source or BUCK files.

We don't use symlinks in the Buck repo today, so let's pre-emptively
forbid them in `.buckconfig`. This allows us to enable a number of
performance improvements in `Parser`.

Test Plan: buck build buck

Reviewed By: sdwilsh

fb-gh-sync-id: ddfc00a
shipit-source-id: ddfc00a